fix: change type of activity level coefficient into migration creation query

+++ src/main/resources/db/migration/V1__create_users_table.sql
@@ -6,5 +6,5 @@ CREATE TABLE users (
     password_salt BYTEA NOT NULL,
     weight NUMERIC(6,3) NOT NULL,
     basal_metabolic_rate SMALLINT NOT NULL,
-    activity_level_coefficient SMALLINT NOT NULL
+    activity_level_coefficient NUMERIC(3, 2) NOT NULL
 );
